requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# LangChain ecosystem (compatible versions)
langchain
langchain-community
langchain-core
langgraph
langchain-google-genai

# Data models and utilities
pydantic==2.7.4
typing-extensions==4.12.2
sqlalchemy==2.0.31

# Database drivers
psycopg2-binary==2.9.9  # PostgreSQL
# pymysql==1.1.0         # MySQL (uncomment if using MySQL)

# Additional utilities
requests==2.31.0
aiofiles==23.2.1

# Optional: for better async support
asyncpg==0.29.0